refactor(api): reclassify the identity providers as domain gateways (#19)
CurrentUserProvider and CurrentTeamProvider were application-layer interfaces
whose only implementations live in infrastructure/identity — driven ports in
everything but name and location. Move them into domain/port as
CurrentUserGateway and CurrentTeamGateway (a flock-accepted port suffix) and
rename their adapters to UserContextCurrentUserAdapter and
UserContextCurrentTeamAdapter, so the classes flock recognises as adapters carry
an adapter name.

Naming them as ports makes flock see them as ports, which surfaces 14
ApiCannotDependOnPorts findings: every controller reads the caller and team ids
straight from these two. That dependency is deliberate — the identity gateways
answer "who is calling", request context the inbound adapter owns, and passing
the ids in as explicit arguments is what keeps the services pure functions of
their parameters. Exempted by exact type name via the rule's own
allowedPortTypes, so no other port slips through: adding SeasonRepository to a
controller still fails, and AuthController's TeamMemberRepository dependency —
a real layering breach — stays baselined.

Behaviour is unchanged; the only baseline edit is the renamed file in the
pre-existing AdapterCannotDependOnAdapter entry.
