+++ api/src/main/kotlin/com/github/zzave/teambalance/api/infrastructure/identity/UserContextCurrentUserAdapter.kt
@@ -1,11 +1,11 @@
 package com.github.zzave.teambalance.api.infrastructure.identity
 
-import com.github.zzave.teambalance.api.application.CurrentUserProvider
+import com.github.zzave.teambalance.api.domain.port.CurrentUserGateway
 import org.springframework.stereotype.Component
 import java.util.UUID
 
 @Component
-class UserContextCurrentUserProvider : CurrentUserProvider {
+class UserContextCurrentUserAdapter : CurrentUserGateway {
     override fun getCurrentUserId(): UUID? = UserContext.get()
     override fun requireCurrentUserId(): UUID = UserContext.require()
 }
