+++ api/src/main/kotlin/com/github/zzave/teambalance/api/interfaces/InvitationController.kt
@@ -1,8 +1,8 @@
 package com.github.zzave.teambalance.api.interfaces
 
-import com.github.zzave.teambalance.api.application.CurrentTeamProvider
-import com.github.zzave.teambalance.api.application.CurrentUserProvider
 import com.github.zzave.teambalance.api.application.InvitationService
+import com.github.zzave.teambalance.api.domain.port.CurrentTeamGateway
+import com.github.zzave.teambalance.api.domain.port.CurrentUserGateway
 import com.github.zzave.teambalance.api.interfaces.generated.endpoint.AcceptInvitation
 import com.github.zzave.teambalance.api.interfaces.generated.endpoint.CreateInvitation
 import com.github.zzave.teambalance.api.interfaces.generated.endpoint.ExpireInvitations
@@ -14,16 +14,16 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 class InvitationController(
     private val invitationService: InvitationService,
-    private val currentUserProvider: CurrentUserProvider,
-    private val currentTeamProvider: CurrentTeamProvider,
+    private val currentUserGateway: CurrentUserGateway,
+    private val currentTeamGateway: CurrentTeamGateway,
 ) : CreateInvitation.Handler,
     AcceptInvitation.Handler,
     ExpireInvitations.Handler,
     RotateInvitation.Handler {
 
     override suspend fun createInvitation(request: CreateInvitation.Request): CreateInvitation.Response<*> {
-        val userId = currentUserProvider.requireCurrentUserId()
-        val teamId = currentTeamProvider.requireCurrentTeamId()
+        val userId = currentUserGateway.requireCurrentUserId()
+        val teamId = currentTeamGateway.requireCurrentTeamId()
 
         val invitation = invitationService.generateInviteLink(callerId = userId, teamId = teamId)
         return CreateInvitation.Response201(
@@ -35,23 +35,23 @@ class InvitationController(
     }
 
     override suspend fun acceptInvitation(request: AcceptInvitation.Request): AcceptInvitation.Response<*> {
-        val userId = currentUserProvider.requireCurrentUserId()
+        val userId = currentUserGateway.requireCurrentUserId()
         val teamId = invitationService.acceptInvitation(request.path.token, userId)
             ?: return AcceptInvitation.Response404(Unit)
         return AcceptInvitation.Response200(AcceptedInvitation(teamId = teamId.toString()))
     }
 
     override suspend fun expireInvitations(request: ExpireInvitations.Request): ExpireInvitations.Response<*> {
-        val userId = currentUserProvider.requireCurrentUserId()
-        val teamId = currentTeamProvider.requireCurrentTeamId()
+        val userId = currentUserGateway.requireCurrentUserId()
+        val teamId = currentTeamGateway.requireCurrentTeamId()
 
         invitationService.expireActiveInvitations(callerId = userId, teamId = teamId)
         return ExpireInvitations.Response204(Unit)
     }
 
     override suspend fun rotateInvitation(request: RotateInvitation.Request): RotateInvitation.Response<*> {
-        val userId = currentUserProvider.requireCurrentUserId()
-        val teamId = currentTeamProvider.requireCurrentTeamId()
+        val userId = currentUserGateway.requireCurrentUserId()
+        val teamId = currentTeamGateway.requireCurrentTeamId()
 
         val invitation = invitationService.rotateInviteLink(callerId = userId, teamId = teamId)
         return RotateInvitation.Response201(
