+++ api/src/main/kotlin/com/github/zzave/teambalance/api/interfaces/MemberController.kt
@@ -1,11 +1,11 @@
 package com.github.zzave.teambalance.api.interfaces
 
 import com.github.zzave.teambalance.api.application.AuthorizationService
-import com.github.zzave.teambalance.api.application.CurrentTeamProvider
-import com.github.zzave.teambalance.api.application.CurrentUserProvider
 import com.github.zzave.teambalance.api.application.MemberService
 import com.github.zzave.teambalance.api.domain.model.Role
 import com.github.zzave.teambalance.api.domain.model.TeamMember
+import com.github.zzave.teambalance.api.domain.port.CurrentTeamGateway
+import com.github.zzave.teambalance.api.domain.port.CurrentUserGateway
 import com.github.zzave.teambalance.api.interfaces.generated.endpoint.CompleteOnboarding
 import com.github.zzave.teambalance.api.interfaces.generated.endpoint.GetCurrentMember
 import com.github.zzave.teambalance.api.interfaces.generated.endpoint.ListMembers
@@ -20,8 +20,8 @@ import java.util.UUID
 @RestController
 class MemberController(
     private val memberService: MemberService,
-    private val currentUserProvider: CurrentUserProvider,
-    private val currentTeamProvider: CurrentTeamProvider,
+    private val currentUserGateway: CurrentUserGateway,
+    private val currentTeamGateway: CurrentTeamGateway,
     private val authorizationService: AuthorizationService,
 ) : GetCurrentMember.Handler,
     ListMembers.Handler,
@@ -30,21 +30,21 @@ class MemberController(
     RemoveMember.Handler {
 
     override suspend fun getCurrentMember(request: GetCurrentMember.Request): GetCurrentMember.Response<*> {
-        val userId = currentUserProvider.requireCurrentUserId()
-        val teamId = currentTeamProvider.requireCurrentTeamId()
+        val userId = currentUserGateway.requireCurrentUserId()
+        val teamId = currentTeamGateway.requireCurrentTeamId()
         return GetCurrentMember.Response200(memberService.getMember(teamId, userId).toDto())
     }
 
     override suspend fun listMembers(request: ListMembers.Request): ListMembers.Response<*> {
-        val caller = currentUserProvider.requireCurrentUserId()
-        val teamId = currentTeamProvider.requireCurrentTeamId()
+        val caller = currentUserGateway.requireCurrentUserId()
+        val teamId = currentTeamGateway.requireCurrentTeamId()
         authorizationService.requireAdmin(caller, teamId)
         return ListMembers.Response200(MemberList(memberService.listMembers(teamId).map { it.toDto() }))
     }
 
     override suspend fun updateMember(request: UpdateMember.Request): UpdateMember.Response<*> {
-        val caller = currentUserProvider.requireCurrentUserId()
-        val teamId = currentTeamProvider.requireCurrentTeamId()
+        val caller = currentUserGateway.requireCurrentUserId()
+        val teamId = currentTeamGateway.requireCurrentTeamId()
         // Admin-vs-self and role guards are enforced in the service, not here.
         val updated = memberService.updateMember(
             callerId = caller,
@@ -58,8 +58,8 @@ class MemberController(
     }
 
     override suspend fun completeOnboarding(request: CompleteOnboarding.Request): CompleteOnboarding.Response<*> {
-        val userId = currentUserProvider.requireCurrentUserId()
-        val teamId = currentTeamProvider.requireCurrentTeamId()
+        val userId = currentUserGateway.requireCurrentUserId()
+        val teamId = currentTeamGateway.requireCurrentTeamId()
         // Onboarding is self-only and never changes role — the request's role field is ignored.
         val updated = memberService.completeOnboarding(
             userId = userId,
@@ -71,8 +71,8 @@ class MemberController(
     }
 
     override suspend fun removeMember(request: RemoveMember.Request): RemoveMember.Response<*> {
-        val caller = currentUserProvider.requireCurrentUserId()
-        val teamId = currentTeamProvider.requireCurrentTeamId()
+        val caller = currentUserGateway.requireCurrentUserId()
+        val teamId = currentTeamGateway.requireCurrentTeamId()
         memberService.removeMember(caller, teamId, UUID.fromString(request.path.userId))
         return RemoveMember.Response204(Unit)
     }
